Moved objects from utils to monitorlib\rid and refactored to move test building out of test_executor and into state replayer
